src/requirements.txt
transformers
flask
sentence-transformers
tensorflow
tf-keras
Pillow
requests
torch